Improvements for library use
